requirements.txt
imageio==2.22.4
numpy==1.23.3
opencv_python==4.6.0.66
Pillow==9.3.0
